requirements.txt
pillow==4.2.1
pygame==1.9.3
python-bidi==0.4.0
peewee==2.10.1